Fixed #10, bug in plot_double_psupertime
plot_double_psupertime (plus a couple of other functions) had a logic error
in handling whether they took double_obj as argument, or took both psuper_1
and psuper_2. This is fixed, plus the documentation is expanded and hopefully
clearer.

Not sure if I've fixed the error in this call, but we'll see...
`plot_double_psupertime(double_psupertime(ps1, ps2))`