Fix #20: replace trendVar with modelGeneVar
Caused by change to `scran`. Note that this makes using the 'span'
parameter tricky, so I've removed it. It's still possible to do this,
but I think it makes more sense to do it outside psupertime, restrict to
the genes you want, and run psupertime with `sel_genes='all'`.

I made a couple of other (mainly cosmetic) changes:
- internal function names now start with a .
- added basic unit tests (even super basic tests are really useful!)